backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
setuptools==75.8.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
wheel==0.45.1
